Don't remove unfinished keyfiles from recovered panics
OOM errors (common when specifying too much memory for Argon2id) are
fatal instead of panicing and there is no stack unwinding or defers
run.  This was the only reason the panic was recovered and re-thrown,
but this technique is useless for OOM.

A better solution to not leave around unfinished keyfiles would be to
call Argon2id first before creating and writing to the files.