Authenticate protocol version before usage
This fixes a potential security issue with the design of the ss
stream.  In version 1 streams, the protocol version, while
authenticated, was negotiated prior to being authenticated.  This
would have made ss susceptible to the "cryptographic doom principle",
had a new version been introduced without fixing the current design.

This change implements a backwards-incompatible fix for this issue.
Existing encrypted streams cannot and will not be decrypted by a new
ss.  The fix is achieved by first decapsulating the symmetric key (now
the first data in the stream), followed by authenticating and
decrypting the sealed protocol version.  If/when later versions are
introduced and additional fields are necessary to be encoded, these
will be sealed in a separate message before the payload begins.