Add symmetric passphrase encryption features
This change adds a -passphrase flag to the encrypt command.  With this
flag, PKI encryption is disabled and the input is encrypted with a key
derived from a prompted passphrase.  Key derivation from the
passphrase is performed with Argon2id.

To support this feature, the stream format has once again changed.
Streaming of encrypted/decrypted chunks has been abstracted away from
the symmetric key handling.  Encrypters are responsible for writing
the correct header for the keying scheme used.  Both encrypters and
decrypters must parse the header to determine the proper keying
scheme, and derive or decrypt the key from the header and user secrets
as needed.  The entire header is ChaCha20-Poly1305 associated data.

Usage of the decrypt command has not been changed, and decrypt
understands both Streamlined NTRU Prime 4591^761 and Argon2id keying
schemes.

The stream format is incompatible with previous versions.