Further reduce pledge promises when possible
This required hoisting some code (such as opening the input/output
files, and looking up the home directory) earlier in the program, so
that more pledges could be removed before parsing dealing with keyfile
or stream parsing.

The refactorings should also make it a little easier to add unveil(2)
support in the future.