Change Decapsulate API to take seeds, not keypairs.
Future KEMs will require the public key material for HKDF info to derive the
shared AEAD key.  The simplest way to support this is to generate both the
public and private keys from the seed as needed.

For legacy compatibility with sntrup4591761, when the keyfile contains the
full private key, it is used as is.  This legacy KEM scheme does not use the
public key in the key derivation.

GenerateKey no longer returns the full serialized private key, and newly
generated identity keys from this point will only record the seed in the
secret keyfile.